+++ sdk/communication/azure-communication-email/azure/communication/email/_generated/operations/_operations.py
@@ -39,7 +39,7 @@ _SERIALIZER = Serializer()
 _SERIALIZER.client_side_validation = False
 
 
-def build_email_get_operation_status_request(operation_id: str, **kwargs: Any) -> HttpRequest:
+def build_email_get_send_result_request(operation_id: str, **kwargs: Any) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
     _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
@@ -111,7 +111,7 @@ class EmailOperations:
         self._deserialize = input_args.pop(0) if input_args else kwargs.pop("deserializer")
 
     @distributed_trace
-    def get_operation_status(self, operation_id: str, **kwargs: Any) -> JSON:
+    def get_send_result(self, operation_id: str, **kwargs: Any) -> JSON:
         """Gets the status of the email send operation.
 
         Gets the status of the email send operation.
@@ -128,16 +128,21 @@ class EmailOperations:
 
                 # response body for status code(s): 200
                 response == {
-                    "id": "str",  # The unique id of the operation. Required.
+                    "id": "str",  # The unique id of the operation. Use a UUID. Required.
                     "status": "str",  # Status of operation. Required. Known values are:
                       "NotStarted", "Running", "Succeeded", "Failed", and "Canceled".
                     "error": {
-                        "code": "str",  # The error code. Required.
-                        "message": "str",  # The error message. Required.
+                        "additionalInfo": [
+                            {
+                                "info": {},  # Optional. The additional info.
+                                "type": "str"  # Optional. The additional info type.
+                            }
+                        ],
+                        "code": "str",  # Optional. The error code.
                         "details": [
                             ...
                         ],
-                        "innererror": ...,
+                        "message": "str",  # Optional. The error message.
                         "target": "str"  # Optional. The error target.
                     }
                 }
@@ -150,7 +155,7 @@ class EmailOperations:
 
         cls = kwargs.pop("cls", None)  # type: ClsType[JSON]
 
-        request = build_email_get_operation_status_request(
+        request = build_email_get_send_result_request(
             operation_id=operation_id,
             api_version=self._config.api_version,
             headers=_headers,
@@ -298,32 +303,32 @@ class EmailOperations:
                     "recipients": {
                         "to": [
                             {
-                                "email": "str",  # Email address. Required.
+                                "address": "str",  # Email address. Required.
                                 "displayName": "str"  # Optional. Email display name.
                             }
                         ],
                         "bcc": [
                             {
-                                "email": "str",  # Email address. Required.
+                                "address": "str",  # Email address. Required.
                                 "displayName": "str"  # Optional. Email display name.
                             }
                         ],
                         "cc": [
                             {
-                                "email": "str",  # Email address. Required.
+                                "address": "str",  # Email address. Required.
                                 "displayName": "str"  # Optional. Email display name.
                             }
                         ]
                     },
-                    "senderEmail": "str",  # Sender email address from a verified domain.
+                    "senderAddress": "str",  # Sender email address from a verified domain.
                       Required.
                     "attachments": [
                         {
                             "contentBytesBase64": "str",  # Base64 encoded contents of
                               the attachment. Required.
-                            "name": "str",  # Name of the attachment. Required.
-                            "type": "str"  # MIME type of the content being attached.
-                              Required.
+                            "contentType": "str",  # MIME type of the content being
+                              attached. Required.
+                            "name": "str"  # Name of the attachment. Required.
                         }
                     ],
                     "disableUserEngagementTracking": bool,  # Optional. Indicates whether user
@@ -334,7 +339,7 @@ class EmailOperations:
                     },
                     "replyTo": [
                         {
-                            "email": "str",  # Email address. Required.
+                            "address": "str",  # Email address. Required.
                             "displayName": "str"  # Optional. Email display name.
                         }
                     ]
@@ -342,16 +347,21 @@ class EmailOperations:
 
                 # response body for status code(s): 202
                 response == {
-                    "id": "str",  # The unique id of the operation. Required.
+                    "id": "str",  # The unique id of the operation. Use a UUID. Required.
                     "status": "str",  # Status of operation. Required. Known values are:
                       "NotStarted", "Running", "Succeeded", "Failed", and "Canceled".
                     "error": {
-                        "code": "str",  # The error code. Required.
-                        "message": "str",  # The error message. Required.
+                        "additionalInfo": [
+                            {
+                                "info": {},  # Optional. The additional info.
+                                "type": "str"  # Optional. The additional info type.
+                            }
+                        ],
+                        "code": "str",  # Optional. The error code.
                         "details": [
                             ...
                         ],
-                        "innererror": ...,
+                        "message": "str",  # Optional. The error message.
                         "target": "str"  # Optional. The error target.
                     }
                 }
@@ -398,16 +408,21 @@ class EmailOperations:
 
                 # response body for status code(s): 202
                 response == {
-                    "id": "str",  # The unique id of the operation. Required.
+                    "id": "str",  # The unique id of the operation. Use a UUID. Required.
                     "status": "str",  # Status of operation. Required. Known values are:
                       "NotStarted", "Running", "Succeeded", "Failed", and "Canceled".
                     "error": {
-                        "code": "str",  # The error code. Required.
-                        "message": "str",  # The error message. Required.
+                        "additionalInfo": [
+                            {
+                                "info": {},  # Optional. The additional info.
+                                "type": "str"  # Optional. The additional info type.
+                            }
+                        ],
+                        "code": "str",  # Optional. The error code.
                         "details": [
                             ...
                         ],
-                        "innererror": ...,
+                        "message": "str",  # Optional. The error message.
                         "target": "str"  # Optional. The error target.
                     }
                 }
@@ -454,16 +469,21 @@ class EmailOperations:
 
                 # response body for status code(s): 202
                 response == {
-                    "id": "str",  # The unique id of the operation. Required.
+                    "id": "str",  # The unique id of the operation. Use a UUID. Required.
                     "status": "str",  # Status of operation. Required. Known values are:
                       "NotStarted", "Running", "Succeeded", "Failed", and "Canceled".
                     "error": {
-                        "code": "str",  # The error code. Required.
-                        "message": "str",  # The error message. Required.
+                        "additionalInfo": [
+                            {
+                                "info": {},  # Optional. The additional info.
+                                "type": "str"  # Optional. The additional info type.
+                            }
+                        ],
+                        "code": "str",  # Optional. The error code.
                         "details": [
                             ...
                         ],
-                        "innererror": ...,
+                        "message": "str",  # Optional. The error message.
                         "target": "str"  # Optional. The error target.
                     }
                 }