+++ sdk/communication/azure-communication-email/samples/send_email_to_single_recipient_sample_async.py
@@ -47,12 +47,12 @@ class EmailSingleRecipientSampleAsync(object):
             "recipients": {
                 "to": [
                     {
-                        "email": self.recipient_address,
+                        "address": self.recipient_address,
                         "displayName": "Customer Name"
                     }
                 ]
             },
-            "senderEmail": self.sender_address
+            "senderAddress": self.sender_address
         }
 
         async with email_client: