+++ sdk/communication/azure-communication-email/samples/send_email_with_attachments_sample.py
@@ -56,16 +56,16 @@ class EmailWithAttachmentSample(object):
             "recipients": {
                 "to": [
                     {
-                        "email": self.recipient_address,
+                        "address": self.recipient_address,
                         "displayName": "Customer Name"
                     }
                 ]
             },
-            "senderEmail": self.sender_address,
+            "senderAddress": self.sender_address,
             "attachments": [
                 {
                     "name": "attachment.txt",
-                    "type": "txt",
+                    "contentType": "text/plain",
                     "contentBytesBase64": file_bytes_b64.decode()
                 }
             ]