+++ sdk/communication/azure-communication-email/tests/test_email_client_e2e_async.py
@@ -24,12 +24,12 @@ class TestEmailClient(AzureRecordedTestCase):
             "recipients": {
                 "to": [
                     {
-                        "email": self.recipient_address,
+                        "address": self.recipient_address,
                         "displayName": "Customer Name"
                     }
                 ]
             },
-            "sender": self.sender_address
+            "senderAddress": self.sender_address
         }
 
         async with email_client:
@@ -50,16 +50,16 @@ class TestEmailClient(AzureRecordedTestCase):
             "recipients": {
                 "to": [
                     {
-                        "email": self.recipient_address,
+                        "address": self.recipient_address,
                         "displayName": "Customer Name"
                     },
                     {
-                        "email": self.recipient_address,
+                        "address": self.recipient_address,
                         "displayName": "Customer Name 2"
                     }
                 ]
             },
-            "sender": self.sender_address
+            "senderAddress": self.sender_address
         }
 
         async with email_client:
@@ -81,16 +81,16 @@ class TestEmailClient(AzureRecordedTestCase):
             "recipients": {
                 "to": [
                     {
-                        "email": self.recipient_address,
+                        "address": self.recipient_address,
                         "displayName": "Customer Name"
                     }
                 ]
             },
-            "sender": self.sender_address,
+            "senderAddress": self.sender_address,
             "attachments": [
                 {
                     "name": "readme.txt",
-                    "type": "txt",
+                    "contentType": "text/plain",
                     "contentBytesBase64": "ZW1haWwgdGVzdCBhdHRhY2htZW50" #cspell:disable-line
                 }
             ]